requirements.txt
openai
python-dotenv
scikit-learn
numpy